Refactor SavedRecipesPage into smaller components: - Extracted CollectionSidebar.tsx for cleaner code. - Fixed layout shifts using absolute positioning for the collection picker. - Implemented global scrollbar hiding and thin border aesthetic. - Removed periods from headers and verified sidebar collection links.
